src/requirements.txt
# MCP SDK
mcp[cli]>=1.6.0

# HTTP requests
requests>=2.32.0

# JSON handling
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0